create utilities module add log_versions
